+++ crates/papyrus_gateway/src/lib.rs
@@ -1,8 +1,5 @@
 mod api;
 mod block;
-pub mod broadcasted_transaction;
-#[cfg(test)]
-mod broadcasted_transaction_test;
 mod gateway_metrics;
 #[cfg(test)]
 mod gateway_test;
