+++ crates/papyrus_gateway/src/v0_3_0/mod.rs
@@ -1,5 +1,8 @@
 pub mod api;
 pub mod block;
+pub mod broadcasted_transaction;
+#[cfg(test)]
+mod broadcasted_transaction_test;
 pub mod deprecated_contract_class;
 #[cfg(test)]
 mod deprecated_contract_class_test;
